Add exception in case no items found in product dashboard

+++ Configurations/TestSettings.cs
@@ -7,6 +7,6 @@
         public string Password { get; set; } = string.Empty;
         public bool Headless { get; set; }
         public required string BaseUrl { get; set; }
-        public float SlowMo { get; set; }
+        public int SlowMo { get; set; }
     }
 }
